ide: fixed to compile
git-svn-id: svn://ultimatepp.org/upp/trunk@13486 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/TextDiffCtrl/PatchDiff.cpp
@@ -128,7 +128,7 @@ bool PatchDiff::Open(const char *patch_path, const Vector<String>& target_dirs)
 int PatchDiff::GetFileIndex() const
 {
 	int ii = files.GetCursor();
-	return ii >= 0 ? files.Get(ii).data : -1;
+	return ii >= 0 ? (int)files.Get(ii).data : -1;
 }
 
 void PatchDiff::File()
